Updated dates handling
Dates don't exist in Rust so convert them to RFC3339 before when deserializing. The return type has no concept of dates.

There is an argument to be made about attempting to parse RFC3339 when serializing strings and seeing if we can convert to a Date but that has two downsides:
1. Cost. Parsing is _not_ free
2. It is done on all strings, regardless of whether it is wanted to be converted to a Date or not. There _should_ be a way to let users convert to dates manually but that is beyond the scope for me